feat(js): basic AnalyzeExpense support for JS/TS
Includes refactor to split out files a bit more.
