+++ src-js/package.json
@@ -32,7 +32,7 @@
   },
   "homepage": "https://github.com/aws-samples/amazon-textract-response-parser#readme",
   "devDependencies": {
-    "@aws-sdk/client-textract": "^3.16.0",
+    "@aws-sdk/client-textract": "^3.24.0",
     "@types/jest": "^26.0.23",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
