+++ src-js/test/unit/document.test.ts
@@ -44,6 +44,9 @@ describe("TextractDocument", () => {
     const firstIteratorPage = [...doc.iterPages()][0];
     const firstPage = doc.pageNumber(1);
     expect(firstIteratorPage).toStrictEqual(firstPage);
+    expect(firstPage.parentDocument).toBe(doc);
+    expect(firstPage.geometry.parentObject).toBe(firstPage);
+
     expect(firstPage.nLines).toStrictEqual(22);
     expect(firstPage.lineAtIndex(0).words.length).toStrictEqual(2);
     expect([...firstPage.iterLines()].reduce((acc, next) => acc + next.words.length, 0)).toStrictEqual(53);
@@ -61,6 +64,23 @@ describe("TextractDocument", () => {
     expect(() => line.wordAtIndex(line.words.length)).toThrow(/Word index/);
   });
 
+  it("should correctly load basic table properties", () => {
+    const doc = new TextractDocument(testResponseJson);
+    expect(doc.pageNumber(1).nTables).toStrictEqual(1);
+
+    const table = doc.pageNumber(1).tableAtIndex(0);
+    expect(table.confidence).toBeGreaterThan(1); // (<1% very unlikely)
+    expect(table.confidence).toBeLessThanOrEqual(100);
+    expect(table.geometry.parentObject).toBe(table);
+  });
+
+  it("should throw errors on out-of-bounds tables", () => {
+    const doc = new TextractDocument(testResponseJson);
+    const page = doc.pageNumber(1);
+    expect(() => page.tableAtIndex(-1)).toThrow(/Table index/);
+    expect(() => page.tableAtIndex(page.nTables)).toThrow(/Table index/);
+  });
+
   it("should correctly load table dimensions", () => {
     const doc = new TextractDocument(testResponseJson);
     expect(doc.pageNumber(1).nTables).toStrictEqual(1);
@@ -68,6 +88,7 @@ describe("TextractDocument", () => {
     const table = doc.pageNumber(1).tableAtIndex(0);
     expect(table.nColumns).toStrictEqual(5);
     expect(table.nRows).toStrictEqual(5);
+    expect(table.nCells).toStrictEqual(25);
   });
 
   it("should index cells by row and column", () => {
@@ -78,6 +99,9 @@ describe("TextractDocument", () => {
 
     expect(table.cellAt(2, 2)?.text).toMatch("End Date");
     expect(table.cellAt(3, 2)?.text).toMatch("6/30/2013");
+    // Strict mode:
+    expect(table.cellAt(2, 2, true)?.text).toMatch("End Date");
+    expect(table.cellAt(3, 2, true)?.text).toMatch("6/30/2013");
   });
 
   it("should fetch row cells by index", () => {
@@ -86,6 +110,8 @@ describe("TextractDocument", () => {
     const table = [...doc.pageNumber(1).iterTables()][0];
 
     expect(table.cellsAt(2, null).length).toStrictEqual(5);
+    // Strict mode:
+    expect(table.cellsAt(2, null, true).length).toStrictEqual(5);
   });
 
   it("should fetch column cells by index", () => {
@@ -94,6 +120,8 @@ describe("TextractDocument", () => {
     const table = [...doc.pageNumber(1).iterTables()][0];
 
     expect(table.cellsAt(null, 2).length).toStrictEqual(5);
+    // Strict mode:
+    expect(table.cellsAt(null, 2, true).length).toStrictEqual(5);
   });
 
   it("should iterate table rows", () => {
@@ -103,6 +131,7 @@ describe("TextractDocument", () => {
     let nRows = 0;
     for (const row of table.iterRows()) {
       ++nRows;
+      expect(row.parentTable).toStrictEqual(table);
     }
     expect(nRows).toStrictEqual(table.nRows);
     expect(nRows).toStrictEqual(5);
